Fix RLS policies and KRS saving issues

+++ scripts/check_student_bills_table.js
@@ -85,7 +85,7 @@ CREATE TABLE IF NOT EXISTS student_bills (
     id UUID DEFAULT gen_random_uuid() PRIMARY KEY,
     student_id UUID NOT NULL REFERENCES profiles(id) ON DELETE CASCADE,
     description TEXT NOT NULL,
-    amount INTEGER NOT NULL DEFAULT 0,
+    amount BIGINT NOT NULL DEFAULT 0,
     status TEXT NOT NULL DEFAULT 'unpaid' CHECK (status IN ('unpaid', 'pending', 'paid')),
     category TEXT,
     semester TEXT,
